modularize commits function to commit.js and import into script.js
