+++ index.html
@@ -88,6 +88,6 @@
     <p class="text-preset-4">Dev Tracker &copy; 2025</p>
     <p class="text-preset-4">Designed and built by Brian J Meinert</p>
   </footer>
-  <script src="./script.js" defer></script>
+  <script type="module" src="./js/script.js" defer></script>
 </body>
 </html>
