+++ js/script.js
@@ -1,0 +1,3 @@
+import { initializeCommits } from './commit.js';
+
+initializeCommits();
